+++ CHANGELOG.md
@@ -29,6 +29,7 @@
 * Deprecate `allow_plain_identifiers` option (#4167)
 * Exception: Add the ability to customize multiple status codes based on the validation exception (#4017)
 * GraphQL: Fix graphql fetching with Elasticsearch (#4217)
+* ApiLoader: Support `_format` resolving (#4292)
 
 ## 2.6.5